+++ Observer/ObserverAppDelegate.m
@@ -220,7 +220,7 @@
             break;
         }
         default:
-            AKRLog(@"Oh No!, Alert View delegate called for an unknown alert view (tag = %d",alertView.tag);
+            AKRLog(@"Oh No!, Alert View delegate called for an unknown alert view (tag = %ld",(long)alertView.tag);
             break;
     }
 }
