+++ Observer/Protocols/ProtocolFeatureAllowedLocations.m
@@ -70,7 +70,7 @@
                     [self defineReadonlyProperties:(NSArray *)json];
                     break;
                 default:
-                    AKRLog(@"Unsupported version (%d) of the NPS-Protocol-Specification", version);
+                    AKRLog(@"Unsupported version (%ld) of the NPS-Protocol-Specification", (long)version);
                     break;
             }
         }
