+++ Observer/Protocols/ProtocolMissionFeature.m
@@ -19,7 +19,7 @@
                     [self defineMissionReadonlyProperties:json version:version];
                     break;
                 default:
-                    AKRLog(@"Unsupported version (%d) of the NPS-Protocol-Specification", version);
+                    AKRLog(@"Unsupported version (%ld) of the NPS-Protocol-Specification", (long)version);
                     break;
             }
         }
