+++ Observer/Surveys/SurveyUploadTableViewController.m
@@ -94,7 +94,7 @@
             break;
         }
         default:
-            AKRLog(@"Oh No!, Alert View delegate called for an unknown alert view (tag = %d",alertView.tag);
+            AKRLog(@"Oh No!, Alert View delegate called for an unknown alert view (tag = %ld",(long)alertView.tag);
             break;
     }
 }
